Fix #76
